backend/requirements.txt
fastapi>=0.100.0
sqlmodel>=0.0.14
uvicorn>=0.22.0
alembic>=1.12.0
psycopg2-binary>=2.9.6
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
pyyaml>=6.0
openai>=1.3.5
